Fixed hanging test

+++ src/Space/Shape/HyperCube.hpp
@@ -216,7 +216,7 @@ public:
 		// Cumulative value
 		size_t val = 0;
 		size_t cum_val = 0;
-		for(size_t i = d - 1; i >= 0 ; i--)
+		for(long int i = d - 1; i >= 0 ; i--)
 		{
 			// check the out-of-bound, outside is assumed to be -1 so  (- pos_n_zero[i+1] - 1) = 0
 			if (i+1 < d)
@@ -224,10 +224,10 @@ public:
 			else
 				val = pos_n_zero[i];
 
-			for (size_t j = 0 ; j < val; j++)
+			for (long int j = 0 ; j < val; j++)
 			{
 				// C is not safe check the limit
-				if (dim-cum_val-j-1 >= 0 && i > 0 && dim-cum_val-j >= i)
+				if (((long int)dim)-cum_val-j-1 >= 0 && i > 0 && dim-cum_val-j >= i)
 					lin_id += openfpm::math::C(dim-cum_val-j-1,i);
 				else
 					lin_id += 1;
